Changed - Moved model classes into subfolder - Moved request-related classes into the Request folder
