+++ dotnet/Xrm.Crm.WebApi.Test/RequestFormatterTest.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using Newtonsoft.Json.Linq;
 using Xrm.Crm.WebApi;
 using Xrm.Crm.WebApi.BatchOperations;
+using Xrm.Crm.WebApi.Models;
 using Xunit;
 using Xrm.Crm.WebApi.Request;
