+++ dotnet/Xrm.Crm.WebApi/Models/ActivityParty.cs
@@ -1,7 +1,7 @@
 using System.Collections.Generic;
 using Xrm.Crm.WebApi.Enums;
 
-namespace Xrm.Crm.WebApi
+namespace Xrm.Crm.WebApi.Models
 {
     public class ActivityParty
     {
