+++ dotnet/Xrm.Crm.WebApi/Request/RequestEntityParser.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using Newtonsoft.Json.Linq;
+using Xrm.Crm.WebApi.Models;
 
 namespace Xrm.Crm.WebApi.Request
 {
