+++ dotnet/Xrm.Crm.WebApi/Response/QualifyLeadResponseFormatter.cs
@@ -1,6 +1,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using Newtonsoft.Json.Linq;
+using Xrm.Crm.WebApi.Models;
 
 namespace Xrm.Crm.WebApi.Response
 {
