+++ dotnet/Xrm.Crm.WebApi/WebApiMetadata.cs
@@ -5,6 +5,7 @@ using System.Net.Http;
 using System.Threading.Tasks;
 using Newtonsoft.Json.Linq;
 using Xrm.Crm.WebApi.Authorization;
+using Xrm.Crm.WebApi.Models;
 using Xrm.Crm.WebApi.Response;
 
 namespace Xrm.Crm.WebApi
